waymo_dataset/requirements.txt
autopep8>=1.5.4
click>=7.1.2
flake8>=3.8.4
isort>=5.6.4
matplotlib>=3.3.2
mypy>=0.921
numpy>=1.19.2
open3d>=0.15.2
opencv-python>=4.5.5.64
pylint>=2.6.2
tensorflow==2.6.0
waymo-open-dataset-tf-2-6-0>=1.4.4